Refactor: Add routes for TOS and logo files
This commit adds new routes to the Server class in the `Server.ts` file. The `/tos` route serves the TOS.html file, and the `/logo.png` route serves the logo.png file from the public assets folder. This refactor improves the organization and accessibility of these files in the application.

Signed-off-by: Scott Gould <greysilly7@gmail.com>