Replace ad-hoc `banInfo` with typed `Ban` entity
When fetching a ban from the past, previously we were fetching the
User's public data, returning that in `user` (unnecessary).

When requesting a ban, all that's needed is the user's ID and other data
related to the ban. The one modification I made to match the schema is
to `delete` the IP property.

Resolves: #1277